Accept CompoundTag for Writing
Now you can directly pass in a non `NBTData`-wrapped `CompoundTag` into the Write module methods! This makes things easier if you aren't working with `NBTData` objects directly, and with the bare NBT JavaScript object instead. Wrote more about this in the same-named issue on the repo!

In the demo test file, I simply wrote the same data to a buffer, using both the `NBT.write()` and `NBTWriter.write()` functions, one using the existing `NBTData` parameter call, and the other with only the `CompoundTag` and options config, no `NBTData` for that one. As you can see in the console, no errors, and the same identical buffer data is written! Noice.

Oh yeah, just to verify the value of the `CompoundTag` a little more strictly, I added a `data !== null` check in there too, since `typeof null` will return `"object"`. I think I read that it's a weird JavaScript legacy thing, kind of interesting. I don't want any of the library parameters to accept `null` as an NBT value! That would be confusing, haha.

Oooh! While writing that last sentence, decided to check this in TypeScript, and I'm very happy to say that this doesn't pass, aaah :)
Scared me for a moment, thinking about this possibly working, since `null` is technically an object in JS:
```ts
const thingo: object = null;
// Type 'null' is not assignable to type 'object'.
```