NBTData Data Typing
Made the `NBTData` `data` property a little looser typed. This allows you to simply read a set of NBT data using `read()`, not have to know the format, edit some of the keys, then re-write back to a buffer using `write()`! Sometimes I find myself doing this when working with JSON, and having `data` typed as `CompoundTag` made using dot notation to edit some of the keys a bit hectic. I want to suggest to the TypeScript team of being able to limit the data of an object interface with the new `satisfies` keyword, because using an index signature doesn't quite work like I want it to here. It would be nice to essentially just have the `data` property be `any`, but throw errors if you try to assign a property value with something that isn't NBT-serializable. It would be awesome if JSON had this too!

The demo here was inspired from the project I made for someone on Reddit the other day, which edits your Bedrock `level.dat` to disable the `hasBeenLoadedInCreative` flag to false. Using a simple dot notation edit and a re-write back to the buffer is nice and slick in this case. I wish it could be typed a little stronger though.