Boolean Primitive Writing
Added support for writing the boolean primitive type! Note that when buffer'd and re-read, it will simply decome a `ByteTag` of either `0` or `1`. This follows the spec, as their is no actual boolean NBT tag type, but rather it's a shorthand for writing a `0` or `1` `ByteTag`. This is currently what happens with reading and writing SNBT in other implementations too, so this can be expected to happen elsewhere. Nothing new in regards to how NBT works, but something to make note of when working with NBT data, writing and reading. Using boolean primitives are the ony asymmetrical tag type, since you don't get the same result when reading, after writing. That's why my `BooleanTag` type isn't strictly just the `boolean` primitive, because it could also be a `ByteTag<0 | 1>`.