CompoundTag Writing Optional Formats
When writing a simple `CompoundTag` object to a buffer using the Write module, you don't have to specify the root name or endianness anymore. The root name will default to `""`, and the endianness will default to big endian. This wasn't an issue before the direct `CompoundTag` writing addition, because the source `NBTData` object always would have the `name` and `endian` properties present, and those will default to `""` and big endian, so there also needed to be defaults for the non-wrapped `CompoundTag` source objects, in case you didn't pass in the options for those.