Improved Compression Streams API Definitions
Rewrote my Compression Streams API type definitions! I've been looking through more TypeScript library code recently, and now I get how the definitions for the built in objects are structured. Seeing how those work now, I decided to rewrite mine to follow that same pattern, of defining an interface for the resulting object from the class after constructed, the class constructor, and the global property to access it on the `globalThis` object. I was previously just using the `class` syntax to define it. While that works exactly the same (to my knowledge, so far), and in a way looks simpler, these are more in line with how the other class-like constructor functions are defined in the standard types libraries, so I think I should follow how those define the constructor functions (Just looked into it, and full ES6 `class` implementations are also defined with this structure too, rather than with the `class` version, so I think this is probably the way to go. I think it's because you can extend `interface` declarations just by redefining the `interface` again, and this seems to be very useful for maintaining library code, and it's many version changes. Very smart!). (Looked into this/thought about it after writing this, it's called Interface Merging)

I also adjusted some of the Compression module's code a little bit, just to break things apart like I've been doing recently, to use more variables in hopes of outlining how things work, a little cleaner.

I started to implement the "Accepting ArrayBufferLike for Reading" bug, rediscovered the `ArrayBuffer` type definition issue (mentioned in the same bug), and that was what kicked off these few Compression module adjustments. Rather than piling everything together, did these first, so they can be out of the way for working on the bug next :)