Accepting ArrayBufferLike for Reading
Woo hoo, got it working!!! Super happy to figure out the typings dilemma. My hack: Redefine `toString()` on `interface ArrayBuffer` as an `as const` string, `"[object ArrayBuffer]"`. Not the prettiest, but also very beautiful. It doesn't add any types for fake property definitions, and it's only inside of NBTify, it can't escape out of the library. Amazing.

I'm probably gonna add this to the issue page to show how I solved it, as it ended up fixing all of the typing issues between `Uint8Array`, which was one of the holdups for implementing this relatively easy thing.

Ok, so down to the actual changes. Now you can simply pass in a raw `ArrayBuffer` or `SharedArrayBuffer` into the Read module, rather than needing it to be wrapped in a `Uint8Array` view. This is super awesome, as I really like the ability to use this with Promise chaining, and without having to make `Uint8Array` objects everywhere. This will be the most useful on the web side of things I feel, like for use with `fetch()`, `Response()`, things like that. And, probably Workers too, come to think of it! Haven't deeply experimented with Workers yet, but not having to manually wrap anything in a `Uint8Array` will be greatly awesome, less implementation details to worry about. It just works!
```ts
import NBT, { NBTData } from "nbtify";

const result: NBTData = await fetch("./level.dat")
  .then(response => response.arrayBuffer())
  .then(NBT.read);
```