Browser SharedArrayBuffer
Tested out the library with the new `ArrayBufferLike` parameter support, and I discovered a scenario with references to `SharedArrayBuffer` that I forgot to account for. When the site is not enabled with `crossOriginIsolated = true`, `SharedArrayBuffer` is simply `undefined`. So when I tried to call `instanceof` on it, when it's `undefined`, it will throw an error for that. Now I check if it's `undefined` before trying to call `instanceof` on it. Sometimes it was showing up, while using `live-server` (Not on my Ubuntubook, but it did on my Mac?), so I hastily just made sure to delete it from the `window` in the demo. Hacky, but sometimes it be that way :)