One Year Old!!!
Big leaps! Going to update the version on NPM fairly soon, so I'm upgrading the package info to show that. I went through all of my closed issues from the last month and a half or so, and they were majorly about 11 different features/improvements/bugfixes, so I incremented the minor release number to show that. There's gonna be a few more things here and there that I'm gonna do, and there were also some changes that I didn't make repo Issues for, so I'm just gonna even it out to a 12 for the minor release number.
https://docs.npmjs.com/about-semantic-versioning#incrementing-semantic-versions-in-published-packages

It has been an unexpected, *very* productive last few weeks working on this project. Not sure why that was exactly, but I have just felt a huge drive to just get everything implemented like how I've wanted it to work (59 commits??). Maybe having a list of all the things I've wanted to implement has helped, I'm not totally sure. I think that had to have helped with that though, there were so many different things to try and keep track of, if I didn't write those down. I think just having a thing to try and grind at a little bit every day has been really relieving and fun, something to get a little closer to every time you work on it. There's a lot that I can build with what I've made here now, and it's almost a bit overwhelming where I should start next. Being able to define Minecraft files using TypeScript definitions is something I'd never thought of doing just a few months ago. Now that this library can leverage that, it's gonna me amazing to work with the data like that, I love it a lot.

And yeah, today a year ago, I made the first commit to the repo! It wasn't a while later that I started to break down the format a bit more. That was in late February, when I started to rewrite NBT.js into a new thing, which turned into this project! So, the actual library is still a few months away from (a year) becoming it's own thing, but today is still a big day :D

It's mindblowing how many things have hopped me along, thanks to working on this project. I feel like I am a TypeScript guy now, and that *definitely* wasn't the case before. I really like ES Modules, for use with Node and the web. The library has gone through so many phases, maybe I should just make it version 2.0.0 instead? Or is that too soon?

I'm both glad that the library is at this point, but also kind of sad that so much is already complete now :(

I think I've done a lot of great work on this, and I think I could use a little step back from it for a few days, my brain is kind of soup from getting everything implemented that quick. A lot of the planning thus far has been to prepare for adding the features in the right order, as it makes sense, and each kind of helps each other along. Now that they're all pretty much there, it's actually time to use the library to build cool shit! (Thanks Theo ping.gg XD) Getting out of the in-progress mindset, and back to the inventive mindset is a bit confusing at first, and I think that's what my brain is stuck on/confused about now. I think you have your initial dream for a project, you outline it, then you build it for the longest time, and then you have the project (Not that it's just done, but yeah. Everything I've wanted to make things is essentially added now). Now that cycle is done, and now I'm in the dream phase for the project I don't really know of yet. I've been trying to work on Minecraft world data parsing, and that has been a side project associated with this one, working together to get each other implemented. Now that this project is essentially stable, I can start using this to work on the possible world conveter, which is in the project dream stage still.

I think that could have been more cohesive of a commit message, but I am exhausted from today for some reason, so my brain is kind of just typing whatever I think. I feel like I could add more, but I don't know what exactly. I am also thinking about the docs side of things too, as it will look sick to have a documentation page that describes how to use the library. That will make things so much nicer! It's gonna be my first *actual* complete explainer XD. I'm gonna take my time on it, it's gonna match how nice the library is too :)

If you read this, you are a champ. I don't know why you're here, but thank you. Theo ping.gg is a great YouTuber I found not too long ago, you should check him out. Super inspiring, and love his attitude towards the craft (programming).

Alright, enough talk here. Gonna go try and build more cool shit!

Brandon :)

Edit: Went and looked at the Issues page, and I think my mind is oatmeal rn because I closed 5 of the 11 Issues only *yesterday*, and the 6th only 3 days ago. Is it just me, or is it a bit overwhelming to mentally keep track of what stage of things are in your project, if you do too much of it at once? That's a holdup for me sometimes, and that's definitely what happening now too, come to think of it. I think it's because those are still on your mind of things to test and look out for, even though they are done now. So your brain is still too full from the last changes before you can think about making more. That perfectly describes how it feel like what happens! I think a possibly fix could also be to just mix up my coding sessions a bit more too, since I tend to just power through for a day, have a work day, then power through some more. Then I take a week or two off from the project, cause my mind is twisted on it still. I think that's where my mind is at for the MCRegionJS project right now too. I tried working on that instead tonight, and I was having a hard time getting started on where I left off.

AAh, these are really nice to write. My silly worry is that having such big commit messages will make the repo bigger, I hope not too bad XD

Ok, good night. I should probably take a little break now haha. Cya.

Edit2: Ok. I think I actually worked on writing that for an hour. Wow!