Source Maps + Side Effects
Did some looking into packaging, bundling, minifying, those sort of things. Gonna try out adding source maps, as it will look nice in DevTools for error positions, and it makes it nice to easily see the actual source for the code right in the browser too, with all of it's TypeScript goodness!

In addition to that exploration, found a super cool site called Bundlephobia, and it gives some super neat stats about NPM packages. After seeing how big some of the other libraries are, namely Prismarine NBT, having a bit extra in size for source maps shouldn't be too crazy. You get a nice developer experience out of it too!
https://bundlephobia.com/

One of their metrics mentioned whether the package is side-effect-free, which means if the package does anything to any code outside of it's own area. An example of that would be a polyfill or something, I think. Since bundlers don't know if there are side effects for a package or not, they assume there are, just to be safe and not break anything. After looking into it more, looks like there's a commonly-followed (not standard in terms of the `package.json`) property called `sideEffects`, and bundlers like Webpack will decide if it can perform tree-shaking on your code, if this property is set to `false`. Since NBTify doesn't have any side effects, I can set it to false too! Now it can enable tree-shaking for NBTify wherever someone may need it. I honestly haven't looked into bundlers yet, but I get the idea behind that, and it's definitely something I want to try working towards from the start! Modules are awesome!
https://sgom.es/posts/2020-06-15-everything-you-never-wanted-to-know-about-side-effects/

Looking into how nbt-ts implemented some of their bundling and packaging, I also noticed that they provide both an ESM and CJS export, which is neat. I kind of want to do that, but I also just want to go all in on ESM too. I don't want to prevent other people from using the library, but I think full-ESM is just too nice not to strive for. It works everywhere now, so I think it's good to try and work towards using it wherever you can! I wish it worked cross-browser in Service Workers, that's one of the last things I'm waiting for on that!