Data Module: Part Two
Moved the NBTData object declaration and related types into it's own new, separate module. The Data Module!

Now that a lot more functionality is in the library, I'm going through and organizing things a bit again.

This makes `index.ts` specifically meant for the library's exports exclusively, and that makes it a bit more straightforward as to what it's task is again.

In this commit, I also reordered the appearances between the Tag and Primitive modules. Since the Tag module is the more high-level of the two, since they are (mostly) just types, I am reordering that one to be listed for when imported and exported by the other modules. Along these same lines, I'm not fully sure where I like the Compression module to be listed, since it's fairly high level in terms of where the library uses it. But, then again, it's also low-level since it deals with the raw buffer data itself, and the users of NBTify shouldn't have to worry about using the Compression module, unless they are doing something custom that requires it. So, maybe it does make sense to list it lower, since it is an implementation detail of the library.

Started rewriting the README docs to add some more info, but realized some changes like this could make the import/export syntax a bit nicer for the user.