End of Buffer Error: Part Three
Did some thinking about the `NBTReader` class outside of my coding time, and I realized that the End of Buffer Error handling can actually be it's own standalone method on the `NBTReader` instance, essentially a mirror to `NBTWriter`'s `#accommodate()` method. In this case, this function isn't extending any buffers, but rather, checking if the one it's reading from will have enough bytes to call the upcoming `DataView` accessor calls. This abstracts the error handling into a single place, rather than handling it individually in each of the `NBTReader` reading methods.