Writer Allocate Method
Redid the `#accommodate()` method on `NBTWriter`, which now has an inspired feel from the new `NBTReader` `#allocate` method.

I remembered that Node has the `Buffer.alloc()` method name, and I've been wanting to rename `#accommodate` for a while now (It didn't feel like a very specific name to me originally), so I took inspiration from that and made the new `NBTReader` method, and the original `NBTWriter` method and renamed them to use that wording instead! Allocating sounds more like what it's doing, to me, and I like that it sticks in your mind better as to what the function call is doing, when it's called.

I tidied up the function body a bit to match my more recent personal code structuring. Some of the functionality is a bit mystical to me still, so I'm going to try reimplementing it after this commit, since I didn't want to break anything with this commit.

https://www.reddit.com/r/ProgrammerHumor/comments/zv2qd5/it_shall_not_be_touched/j1o1zv4/?context=8&depth=9 (Reminded me of a post on Reddit I found about this, I find myself trying to do it more and more! Cool to hear that other people do it that way too.)