Compression Format Typings
With the Compression Streams Typings update a few commits ago, I didn't realize that the type exported by the Compression module wouldn't be accessible outside the scope of the library. This makes it so the type is accessible again, both inside and outside the library. If you define a custom type or interface in your `global.d.ts` file, you won't be able to export it in any of your type definitions. I didn't really put that together for some reason, it makes sense now though!