TypedArray Tag Lengths
While implmenting the next feature, I found a hidden issue that I hadn't discovered yet. Turns out I was using the wrong length property when reading and writing to and from the different TypedArray tags, like `ByteArrayTag`, `IntArrayTag`, and `LongArrayTag`. I was using the `byteLength` property before, and this was causing the raw NBT output to encode the length of the array to be much longer than it actually was. Instead, the `length` property is the one to use in this case.

I hadn't discovered this until I started trying to parse Java Edition region and chunk files, where those have multiple instances of using `IntArrayTag` and `LongArrayTag`. Doing the checks against the use of `ByteArrayTag` before were passing, because the size of each byte is only one, so that means `byteLength` and `length` happened to have the same values. For the other TypedArray types though, those have a longer byte length per entry, so that means the resulting length that I was using was *x* as many bytes longer than the real length, *x* being how many bytes each entry is. So, say for `LongArrayTag`, that would be 8 bytes for each entry, vs 1.