Deflate API Option
Moved the NBT API `"zlib"` compression option over to be `"deflate"` instead, so now it is synonymous with the format that it represents.

Note that the `hasZlibHeader()` function is still aptly named, as it looks like Deflate is the format you are compressing with, and ZLIB is what puts the header there, and that's it's own format, unrelated to the Deflate implmentation, I think. Of course the header does describe how the data was compressed, but it's not specifically part of the compressed data payload format.