Class Compounds + Object Errors: Part Two
This makes the accepted types into the Write API a bit more lenient, so now you can pass classes in, and it won't throw a type error.

This does result in having less type-checking of the values going into the API, but it will allow you to pass in more things than you could before.

This is how the JSON module handles things, so I'm going to follow in it's footsteps there to make things more serializable than there were before. I want to be able to serialize the same things that are serializale by `JSON.stringify()`, which means you could go back and forth from NBT without issue! (Kind of, if you aren't using any NBT-specific primitives)