Empty Lists: Part Three / Object Errors: Part Four
Expanding off of my findings from the Empty Lists updates a few weeks ago, this is going to adjust the unexpected value handling for Compound and List tags to handle all unexpected value types accordingly.

Rather than throwing an error during the writing process if an unexpected value is encountered, it will simply skip that value or key/value, and move on to the rest of the entries.

To make this work, I am changing back how the `getTagType()` function works again. Now, instead of throwing an error for unsupported NBT primitives, it will instead return either the Tag ID (if the entry is indeed a supported NBT primitive, no change here), and now it will also return `-1` if it's not supported. This allows for more direct value checking, rather than handling the errors as they come up.