Class Compounds + API Type Handling
Been working on this for about the past two hours, and I can't get it to work for the life of me!

This is the closest post that I've found to what I think I'm trying to accomplish.
https://stackoverflow.com/questions/73456021/how-to-restrict-the-allowed-properties-of-a-class-in-typescript

So I want to be able to accept both plain objects, and class-generated ones, and get type error support to verify that the incoming tags are indeed valid, at least from the type surface. I don't like having `any` as the parameter for writing to an NBT buffer, as I have already found out how to validate the property values against plain objects. This simply just isn't working for classes, because those appear to require a symmetrical index signature to the one that the parameter is validating with.

I tried out the `toNBT` Symbol idea I had mentioned in the Class Compounds issue, and it worked fairly nicely. But doing that isn't in-line with the new relaxed object/class error handling that I am trying to work on now.

This article I found might be a very close way to being able to do what I'm looking for. It still has thrown some odd errors relating to index signatures unfortunately though, so I'm not sure, maybe I just wasn't testing it right. I think using a Mapped type may be a nice solution for this, I'm going to continue trying that out to see if I can get the type checking working for *both* plain objects, and ones generated from classes too.

I've worked hard to make all of this strictly typed, I'm not gonna lose that now!