Compound Class Parameter Passing
This is the final way that I have gotten to that will allow both `CompoundTag` satisfying objects, and objects generated from classes.

This removes the type checking specifically from the `read()` and `write()` functions, and the `NBTData` constructor. But, if supporting more class types is the goal, then I think this is the only way to get this working.

This whole headache is because an object literal can validate against an interface/type with an index signature as a valid type, but another `interface` or `class` is not able to validate against the same thing, because it's not strict enough.

The issue code snippet isn't exactly the same thing I'm running into here, but it seems related to it. Ryan Cavanaugh's explanation for it seems to make sense for my situation too, so that could be why it's how it works.
https://github.com/microsoft/TypeScript/issues/15300#issuecomment-332366024

I also tried switching out `CompoundTag` to be a `type` declaration, rather than an `interface` one, but that didn't help with the validation. It also didn't throw a recursive type error, which I was surprised about. I thought I made it an `interface` declaration on purpose because of that. Turns out, only `ListTag` has to be an `interface`, because of the recursive type references.

Edit: Related to this, and it was the post that helped me track down the GitHub issue.
https://stackoverflow.com/questions/60697214/how-to-fix-index-signature-is-missing-in-type-error

Edit 2: This is the next test that I did, yet again.

It validates an object literal against the index signature type, which indeed does validate.

In the event that you want to pass a `class` into the data parameter, it must either pass the index signature type, or specify the "toNBT" `Symbol` on the object, which will essentially force the `class` to validate with the parameter type. While this is not ideal, it does keep the object literal type checking working, which is fantastic! The downside is that any `class` (or, `interface` for that matter) can be made valid with the presence of that `Symbol` property. But, say if I were to use the `object` type for the parameter validation (I'm rethinking this commit already XD), then anything that is a child of `Object` (this would allow functions to pass, which I don't like) would pass the parameter check. This is more specific than `any`, which allows primitives to be the top-level parameter (also not supported by NBT), so that doesn't make any more sense than this to allow into the parameter.

After looking into the Stack Overflow article above, the best shining light to this is to use the spread syntax on the `class`, making a new object literal out of it's `class` members. This does have a run-time effect that I don't want to have, but it's both optional, and it actually does exactly what I'm looking for! You get the class validation right in the parameter, you don't have to add an index signature to the `class` definition, and the object literal type checking (for regular, non-spreaded objects) still works perfectly.

```ts
export declare const symm: unique symbol;

export type thingos = boolean | number | string | thingo | thingos[];

export type thingo = {
  [name: string]: thingos;
};

export type hoya = { [symm]: boolean; };

export type checker = thingo | hoya;

export declare function nocie(epic: checker): void;

nocie({
  noice: [4,4,4,4,4],
  eya: /gg/
})

class Thingo implements hoya {
  // [symm] = true;

  Noice = true;
  Heya = /aa/;
}

nocie({ ...new Thingo() });

nocie(/gg/);
```

So glad I read into that post, it appears to be a possible resolution for the time being, until TypeScript supports something nicer than having to do this.

I think we should make a way to define an index signature that is a `satisfies` block, so you could define it only to type check whatever value you are type checking it against.

This seems like a really neat way of doing it actually! Here's a code snippet of what it could look like:

```ts
interface NoiceThing {
  [key: string]: string;
}

// Index signature only for type checking
interface NoiceThingSatisfies {
  satisfies [key: string]: string;
}

// Satisfy a class against an index signature
class Noice satisfies NoiceThing {
  hey = "";
}

declare function noice(thingo: NoiceThing): unknown;

const thing = new Noice() satisfies NoiceThing;

noice(thing);
```

Ok, actually gonna commit this now, then gonna start using the spread method when making my `CompoundTag` `class` constructors. Or, if I don't want the run-time cost of it, I will probably do `as unknown as CompoundTag`, once I know that the `class` is a valid NBT structure. And a comment there for that nonsense too.

```ts
type Tag = boolean | number | string | Tag[] | CompoundTag;

interface CompoundTag {
  [name: string]: Tag;
}

declare function write(data: CompoundTag): Promise<Uint8Array>;

write({ "I'm a valid NBT object": true });

class NonPassing {
  ShouldPass = true;
}

const nonPassing = new NonPassing();

write(nonPassing);

write({ ...nonPassing });

write(nonPassing as unknown as CompoundTag);
```