Compound Class Object Spreading
Alright, this is what I'm going with for now! I tried it out in my test script, and it seems to be working as expected! I didn't even think about that it will work great with `allowJS` and `checkJS`, that's a great bonus! You unfortunately get some run-time effects, from needing to spread the class `{ ...myClass }`, but at least it works like intended in *both* full TypeScript, and type-checked JavaScript.

The ultimate fix for this will be to be able to add `satisfies` declarations to index signatures, and class definitions, or something like that. Even if it's not that, there should be a built-in TypeScript/JSDoc way to tell TSC to check the type like how it is doing it for the spread operator, as that appears to be the main issue here. It's checking the resulting object like a `class` (understandable), but I want it to be checked like an object literal would be. So, another weird idea for something like that could be an `as literal` modifier, maybe.

```ts
interface MyInterface {
  [name: string]: string;
}

const myLiteral: MyInterface = {
  noice: "this-is-valid"
};

class MyClass {
  Noice = "This won't validate, even though the object is valid";
}

// this would work like `as const`. It would type check the class object like it does for an object literal.
const myClass = new MyClass() as literal;
```