List Item Type Coalescing
Constrasting to the original nature of the description for this issue, I am choosing to add error handling for inconsistent array item types, rather than coalescing the values to match the root item tag type for the array (inferred from the first valid element in the array).

This commit is a rework of the array item type checking logic, with now takes use of the new `getTagType()` API, which more explicitly returns whether a given value is a supported tag or not.