Error Wording Adjustments
Oh yeah, there was a link that I was looking at that made me think about the Circular Data Structure Handling bug report. It was the reference to the "shallow copy" that `Array.prototype.filter()` does when acting on an array. I looked into the documentation about that, and that brought me to looking into circular references, and back to how `JSON.stringify()` responds to it.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter

Been using my Chromebook a lot, and I'm starting to think about upgrading to the Framework DIY Edition here, pretty soon. I think the battery on here is starting to age now, or maybe it's just because I spend hours on here, haha. Either way, I think a new daily driver could allow for more headroom with my coding tasks, which would be nice to continue expanding with. Gonna start looking into it a bit more :)