SNBT Indentation
Wow! Been working on this for a few hours now. Had it almost working a few times along the way, so I would restart with what I had learned to make it right to get even closer to having everything indented correctly.

This looks great! There's only a little more formatting I want to add to non-structure-based tags, so primitive-based arrays will still have a little bit of spacing, but not be multi-lined.

This is a super huge benefit to having my own object parser! I would have loved to have these kinds of checks for my JSON formatting in STE back in the day. The ability to format arrays differently, depending on what is in each array, is super awesome. I want the TypedArray tags to be single-lined, since they tend to be really long, the primitive item arrays to be singled lined (string arrays may be one I might make multi-line though), and the arrays which *hold* either ListTag, CompoundTag, or the TypedArray tags will be multi-lined. So the only one left to implement from those, which isn't already added to match those are the possible string arrays, and pretty but single lined, primitive arrays. I think I will keep the TypedArray tags flattened to not have any whitespace. I can look into seeing if it looks nice though :)

Listened to Blackwater Park for the second time tonight. I love it a ton! It was musically easy to comprehend this time around. Gonna listen to that some more this week, and also gonna go further into Mike Keneally's catalog. Listened to Incubus Science today, did not expect that! It sounded very Keneally-ish, completely different from their sound on later albums. It was sick! Glad I mentioned that right now to myself, now I'm gonna make it a plan to listen to it again tomorrow! And Blackwater Park :8