Abstracting Stringify + String.repeat()
Trying a few attempts to make the stringify API to look and work like the rest of my API for NBTify, like working with NBTData objects, and only exposing  parameters that are intended to be configurable by the user.

I want to make the Stringify function flow structured like the Write module, to where there is a base class that handles the internal logic, and the main Stringify function simply calls that to make everything run. Been having trouble moving this current structure over to that syntax, so I'm going to move things in smaller steps before trying to do that. I think if I change some of the other parts of the code first, it will help make it more obvious as how to move things around to also be in that structure.

Looked into a better method of repeating a string segment for a given number of times, and turns out there's a simple method for that, `String.prototype.repeat()`! My workaround before was very hacky, and I went with that since I knew it would work initially (`"".padStart(snippet.length * n, snippet)`).