Function Flattening
Starting to break up the existing general-use functions and moving their implementations into where they are directly used, since some of them are only used in one or two places.

Found a bug where I was asserting the type of a tag to `bigint`, while it actually needed to fully be called with `BigInt()`.

Added some more key types to check with my demo object, more things to be able to check in one file!

Back to the first part, since I only want some of the list tag types to have the formatting, I moved the fancy formatting logic into the `ListTag` handler specifically, since I want the `TypedArray`-based ones to simply be formatted lightly, not with indentation or anything. A few extra `.join()` methods are a bit more explicit as to what each function does too, which I like. It lets each tag have more unique formatting too, since you don't have to pipe everything through the same single function for it. That's the same thing I'm doing for the Read and Write modules; each class has unique methods that manipulate the data directly, and while each of them has fairly similar logic, I let it repeat a little bit so you can more easily see what it's doing to actually work!

In trying to move this logic over to the class-based/function-wrapped setup like the other two modules, I want to try moving over the methods to perform the transformations on a single set of data, rather than returning transformations with each call. So, say for the hypothetical `#writeByte()` function, I want it to append the new formatted SNBT value directly to the "text stream" result thing that it's working on, instead of returning the `${value}b` string itself. That's the model the other two classes use, so while I do like the return-based nature of the function calls, it would be good to have them all work the same way, like a tap feeder running over the data, tag by tag, mapping the old data to a new equivalent mirror, just in the other format.